\name{pwn-package}
\alias{pwn-package}
\alias{pwn}
\docType{package}
\title{
Own your data.
}
\description{
A collection of utility functions for data cleaning, and just getting to know the data better. 
}
\details{
\tabular{ll}{
Package: \tab pwn\cr
Type: \tab Package\cr
Version: \tab 0.1.5\cr
Date: \tab 2015-12-16\cr
License: \tab GPLv3\cr
}
This package is meant to provide a collection of tools to reduce the general agony of data cleaning. 
Everytime you need to work with data from more than one source (or even then), you'll probably have to do it. 
Most of the functions here are meant to be applied directly to your data, and merely require you to supply the name of the dataframe(s) and/or variable in question. 
Functions that were meant to be applied to a vector will require the name of the vector, specified as a string. 
}
\author{
hellhound
Maintainer: hellhound <hellhound007@gmail.com>
}
\references{
no references here. 
}
\keyword{ data }
\seealso{
Nothing to see here. 
}
\examples{
# refer to help file for specific functions. 
}
